Make Github recognize the license
This patch changes the formatting of the ical4j license to the default
formatting for the 3-Clause-BSD license used by Github. This causes
Github to recognize the license and display license information for it.

Note that this changes only the formatting of the license with two
exceptions:

- Neither the name of Ben Fortuna nor the names of any other contributors may be used
+ Neither the name of the copyright holder nor the names of its contributors may be used

- IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
+ IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE

I hope that these changes are ok.

The reasoning for this pull request is that I was approached by by one
of the Apereo Foundation projects to check if this custom license is ok
for the project to use. Knowing that it's just a 3-clause BSD license
would make such questions way easier for developers to decise on their
own ;-)