Change expectation of RecurSpec for intervals
https://www.rfc-editor.org/rfc/rfc5545#section-3.8.5.3 says:

```
The "DTSTART" property
value SHOULD be synchronized with the recurrence rule, if
specified.  The recurrence set generated with a "DTSTART" property
value not synchronized with the recurrence rule is undefined.
```

... so pick a start date that is on a Sunday.

This way the behavior of this test also no longer depends
on on which day of the week the week starts.

Fixes #727