Fixate timezone in P1D test
The reason this test worked in most timezones is because:

* The test specifies input and expected output in local (floating) dates
* The function being tested, `TemporalAmountAdapter.getTime`, produces a `java.util.Date`.
* This means it is forced to make assumptions about the timezone, because `java.util.Date` is timezone-aware. `getTime` converts the `DateTime` to a UTC `Instant` assuming it is in the default timezone, adds a day, and then converts it to a `java.util.Date` in the default timezone.
* Before checking the result is as expected, it is converted back to a local (floating) `DateTime` again, keeping the default timezone.

This fails for the `Europe/Amsterdam` timezone, because in Europe/Amsterdam, the night between 20110326 and 20110327 is exactly the rollover from CET to CEST, leading to an hour difference compared to what you'd expect.

I guess ideally `getTime` should add the duration on the 'floating' time, before converting to `Date`. However, `getTime` is currently deprecated, and the javadoc recommends to "`Instant.from(getDuration().addTo(start))`, where `start` is a `Temporal`". I think it's reasonable to keep `getTime` as-is. The recommended code has the same behavior, but there going via `Instant` is explicit, so it's reasonable to consider this the programmers' responsibility.

This commit describes the behaviour explicitly in the test, and removes
the `Instant.from` from the recommended alternative to `getTime`.

Fixes #731